refactor(model-registry): fold src/models/ (discovery layer) into model_registry/
Closes the audit smell codex flagged in #1414 / earlier scan:
'models/ AND model_registry/ separate dirs. models/mod.rs is the
only file in models/ but it's still actively imported ... Old
source-of-truth ghost that should be deleted but persists.'

Per Joel's "we are not in some maintenance mode — zero users,
FULLBLOWN rust driven dev" calibration: no transition period, no
deprecation shim. Move the file to its new home, fix the one
import site, delete the empty dir.

## What changed

- `src/models/mod.rs` (352 LOC, model-discovery HTTP layer) →
  `src/model_registry/discovery.rs` (single move, no content
  change — git rename detected)
- `src/lib.rs`: remove `pub mod models;` (one line)
- `src/model_registry/mod.rs`: add `pub mod discovery;`
- `src/modules/models.rs`:
  `use crate::models::{ProviderConfig, discover_all};` →
  `use crate::model_registry::discovery::{discover_all, ProviderConfig};`

## Why this is conceptually right

`models/` was the live-discovery layer (queries OpenAI / Groq /
Together API endpoints for available models). `model_registry/`
is the curated static catalog (Rust code post-#1424).

Both are model-metadata responsibilities; both belong under
ONE umbrella. The names diverged historically — `models/` predates
`model_registry/`. Lane A's #1424 made `model_registry/` the
canonical Rust-truth home; this PR closes the loop by pulling
discovery into the same module tree.

The discovery vs catalog distinction is preserved (it's two
sub-modules now, not two top-level dirs), so callers that need
"live API discovery" still get the explicit `discovery::`
namespace.

## Validation

  - `cargo build --features metal,accelerate --lib` clean (no
    errors, no new warnings)
  - `cargo test --features metal,accelerate --lib model_registry`
    16 tests pass
