+++ src/workers/continuum-core/src/model_registry/discovery.rs
